+++ projects/ngx-opendesign/src/public-api.ts
@@ -38,6 +38,7 @@ export * from './lib/layout/stack';
 export * from './lib/layout/grid';
 export * from './lib/layout/layout';
 export * from './lib/table/table';
+export * from './lib/table/table-container';
 export * from './lib/table/sort-header';
 export * from './lib/table/freeze';
 export * from './lib/pagination/pagination';